requirements.txt
certifi==2022.12.7
charset-normalizer==3.3.2
click==7.1.2
cycler==0.11.0
daal==2023.2.1
daal4py==2023.2.1
filelock==3.12.2
fonttools==4.38.0
fsspec==2023.1.0
fuzzy-c-means==1.6.3
fvcore==0.1.5.post20221221
h5py==3.8.0
huggingface-hub==0.16.4
idna==3.6
importlib-metadata==6.7.0
iopath==0.1.10
joblib==1.3.2
kiwisolver==1.4.5
mat73==0.62
matplotlib==3.5.3
numpy==1.21.6
onnx==1.14.1
opencv-python==4.9.0.80
packaging==24.0
pandas==1.3.5
Pillow==9.5.0  
pip==22.3.1
portalocker==2.7.0
protobuf==4.24.4
psutil==5.9.8
py-cpuinfo==9.0.0
pydantic==1.10.15
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
scikit-learn==1.0.2
scikit-learn-intelex==2023.2.1
scipy==1.7.3
seaborn==0.12.2
setuptools==65.6.3
six==1.16.0
tabulate==0.8.10
tbb==2021.11.0
termcolor==2.3.0
thop==0.1.1.post2209072238
threadpoolctl==3.1.0
torch==2.2.2
torchaudio==2.2.2
torchvision==0.17.2
tqdm==4.66.2
typer==0.3.2
typing_extensions==4.7.1
ultralytics==8.0.145
urllib3==2.0.7
wheel==0.38.4
yacs==0.1.8
zipp==3.15.0
